requirements.txt
django
bcrypt
pillow
django-registration-redux
PyMySQL
mysqlclient=1.3.14